WIP compute of telemetry details in the background
